Adds cif compatibility

+++ pdbtools/download.py
@@ -21,7 +21,12 @@ DIRECTORY="/pub/databases/pdb/data/structures/all/pdb/"
 PREFIX="pdb"
 SUFFIX=".ent.gz"
 
-def unZip(some_file,some_output):
+# Added constants for mmCIF
+CIF_DIRECTORY = "/pub/databases/pdb/data/structures/all/mmCIF/"
+CIF_PREFIX = ""
+CIF_SUFFIX = ".cif.gz"
+
+def unZip(some_file, some_output):
     """
     Unzip some_file using the gzip library and write to some_output.
     """
@@ -37,10 +42,19 @@ def unZip(some_file,some_output):
     os.remove(some_file)
 
 def pdbDownload(file_list,hostname=HOSTNAME,directory=DIRECTORY,prefix=PREFIX,
-                suffix=SUFFIX):
+                suffix=SUFFIX, file_format="pdb"):
     """
-    Download all pdb files in file_list and unzip them.
+    Download all PDB or mmCIF files in file_list and unzip them.
+    
+    By default, downloads PDB files from EBI (file_format="pdb").
+    To download mmCIF instead, call with file_format="cif".
     """
+
+    # If user wants CIF, override defaults
+    if file_format.lower() == "cif":
+        directory = CIF_DIRECTORY
+        prefix = CIF_PREFIX
+        suffix = CIF_SUFFIX
 
     success = True
 
@@ -50,12 +64,17 @@ def pdbDownload(file_list,hostname=HOSTNAME,directory=DIRECTORY,prefix=PREFIX,
     ftp.connect(hostname)
     ftp.login()
 
-    # Remove .pdb extensions from file_list
+    # Remove .pdb or .cif extensions from file_list
     for file_index, file in enumerate(file_list):
-        try:
-            file_list[file_index] = file[:file.index(".pdb")]
-        except ValueError:
-            pass
+        for ext in (".pdb", ".cif"):
+            if file.endswith(ext):
+                file_list[file_index] = file_list[file_index][:file.index(ext)]
+
+    # Decide final extension based on file_format
+    if file_format.lower() == "cif":
+        final_ext = ".cif"
+    else:
+        final_ext = ".pdb"
 
     # Download all files in file_list
     to_get = ["%s/%s%s%s" % (directory,prefix,f,suffix) for f in file_list]
@@ -63,18 +82,16 @@ def pdbDownload(file_list,hostname=HOSTNAME,directory=DIRECTORY,prefix=PREFIX,
     for i in range(len(to_get)):
         try:
             ftp.retrbinary("RETR %s" % to_get[i],open(to_write[i],"wb").write)
-            final_name = "%s.pdb" % to_write[i][:to_write[i].index(".")]
+            final_name = "%s%s" % (to_write[i][:to_write[i].index(".")],final_ext)
             unZip(to_write[i],final_name)
             print("%s retrieved successfully." % final_name)
         except ftplib.error_perm:
-            os.remove(to_write[i])
+            if os.path.exists(to_write[i]):
+                os.remove(to_write[i])
             print("ERROR!  %s could not be retrieved!" % file_list[i])
             success = False
 
     # Log out
     ftp.quit()
 
-    if success:
-        return True
-    else:
-        return False
+    return success
